refactor(daemon): relocate generic emit/think helpers into runtime
Prerequisite for moving the per-arch SpecEmit emitters out of the daemon
example: an example cannot be imported by the arch crates that depend on
hipfire-runtime, so the generic helpers the emitters need must live in a
library crate.

- New hipfire-runtime::emit_text with currently_in_think +
  extract_tool_calls_from_text (+ name/args/complete fallbacks) — the
  model-independent ChatML `<think>` / `<tool_call>` parsing the daemon
  used inline. Moved verbatim; daemon imports them (bare call sites
  unchanged), test call-sites point at the runtime path.

- ThinkMode (reasoning-effort level) moves to hipfire-runtime::prompt_frame
  alongside AssistantPrefix. It is a model-independent request parameter;
  the arch-specific frame mapping stays in the consuming arch. This lets
  the upcoming neutral SpecEmitCtx carry it.

Pure relocation, behavior-identical: coherence battery 11/11 OK on
gfx1100 (tool-call row emits the correct <tool_call> JSON via the moved
extractor).
